requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
requests>=2.31.0
pydantic>=2.6.0
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
jinja2>=3.1.2
python-multipart>=0.0.6
aiofiles>=23.2.1
beautifulsoup4>=4.12.0
lxml>=4.9.0